Move navigation components to 'client/src/components/shared/navigation' package
